+++ container-engine-lib/lib/backend_impls/metrics_reporting/metrics_reporting_kurtosis_backend.go
@@ -3,7 +3,8 @@ package metrics_reporting
 import (
 	"context"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface"
-	engine2 "github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/engine"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/engine"
 	"github.com/kurtosis-tech/stacktrace"
 )
 
@@ -16,7 +17,7 @@ func NewMetricsReportingKurtosisBackend(underlying backend_interface.KurtosisBac
 	return &MetricsReportingKurtosisBackend{underlying: underlying}
 }
 
-func (backend *MetricsReportingKurtosisBackend) CreateEngine(ctx context.Context, imageOrgAndRepo string, imageVersionTag string, grpcPortNum uint16, grpcProxyPortNum uint16, engineDataDirpathOnHostMachine string, envVars map[string]string) (*engine2.Engine, error) {
+func (backend *MetricsReportingKurtosisBackend) CreateEngine(ctx context.Context, imageOrgAndRepo string, imageVersionTag string, grpcPortNum uint16, grpcProxyPortNum uint16, engineDataDirpathOnHostMachine string, envVars map[string]string) (*engine.Engine, error) {
 	result, err := backend.underlying.CreateEngine(ctx, imageOrgAndRepo, imageVersionTag, grpcPortNum, grpcProxyPortNum, engineDataDirpathOnHostMachine, envVars)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred creating the engine using image '%v' with tag '%v'", imageOrgAndRepo, imageVersionTag)
@@ -25,7 +26,7 @@ func (backend *MetricsReportingKurtosisBackend) CreateEngine(ctx context.Context
 }
 
 // Gets point-in-time data about engines matching the given filters
-func (backend *MetricsReportingKurtosisBackend) GetEngines(ctx context.Context, filters *engine2.EngineFilters) (map[string]*engine2.Engine, error) {
+func (backend *MetricsReportingKurtosisBackend) GetEngines(ctx context.Context, filters *engine.EngineFilters) (map[string]*engine.Engine, error) {
 	engines, err := backend.underlying.GetEngines(ctx, filters)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred getting engines using filters: %+v", filters)
@@ -33,7 +34,7 @@ func (backend *MetricsReportingKurtosisBackend) GetEngines(ctx context.Context, 
 	return engines, nil
 }
 
-func (backend *MetricsReportingKurtosisBackend) StopEngines(ctx context.Context, filters *engine2.EngineFilters) (
+func (backend *MetricsReportingKurtosisBackend) StopEngines(ctx context.Context, filters *engine.EngineFilters) (
 	successfulIds map[string]bool,
 	failedIds map[string]error,
 	resultErr error,
@@ -45,7 +46,7 @@ func (backend *MetricsReportingKurtosisBackend) StopEngines(ctx context.Context,
 	return successes, failures, nil
 }
 
-func (backend *MetricsReportingKurtosisBackend) DestroyEngines(ctx context.Context, filters *engine2.EngineFilters) (
+func (backend *MetricsReportingKurtosisBackend) DestroyEngines(ctx context.Context, filters *engine.EngineFilters) (
 	successfulIds map[string]bool,
 	failedIds map[string]error,
 	resultErr error,
@@ -53,6 +54,38 @@ func (backend *MetricsReportingKurtosisBackend) DestroyEngines(ctx context.Conte
 	successes, failures, err := backend.underlying.DestroyEngines(ctx, filters)
 	if err != nil {
 		return nil, nil, stacktrace.Propagate(err, "An error occurred destroying engines using filters: %+v", filters)
+	}
+	return successes, failures, nil
+}
+
+func (backend *MetricsReportingKurtosisBackend) CreateEnclave(ctx context.Context, enclaveId string) (*enclave.Enclave, error) {
+	result, err := backend.underlying.CreateEnclave(ctx, enclaveId)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred creating enclave with ID '%v'", enclaveId)
+	}
+	return result, nil
+}
+
+func (backend *MetricsReportingKurtosisBackend) GetEnclaves(ctx context.Context, filters *enclave.EnclaveFilters) (map[string]*enclave.Enclave, error) {
+	results, err := backend.underlying.GetEnclaves(ctx, filters)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred getting enclaves using filters: %+v", filters)
+	}
+	return results, nil
+}
+
+func (backend *MetricsReportingKurtosisBackend) StopEnclaves(ctx context.Context, filters *enclave.EnclaveFilters) (successfulEnclaveIds map[string]bool, erroredEnclaveIds map[string]error, resultErr error) {
+	successes, failures, err := backend.underlying.StopEnclaves(ctx, filters)
+	if err != nil {
+		return nil, nil, stacktrace.Propagate(err, "An error occurred stopping enclaves using filters: %+v", filters)
+	}
+	return successes, failures, nil
+}
+
+func (backend *MetricsReportingKurtosisBackend) DestroyEnclaves(ctx context.Context, filters *enclave.EnclaveFilters) (successfulEnclaveIds map[string]bool, erroredEnclaveIds map[string]error, resultErr error) {
+	successes, failures, err := backend.underlying.DestroyEnclaves(ctx, filters)
+	if err != nil {
+		return nil, nil, stacktrace.Propagate(err, "An error occurred destroying enclaves using filters: %+v", filters)
 	}
 	return successes, failures, nil
 }